Finalização de testes

+++ api/src/main/java/com/ecommerce/dataprovider/OrderDataProvider.java
@@ -1,6 +1,7 @@
 package com.ecommerce.dataprovider;
 
 import com.ecommerce.core.handler.exception.HandlerValidationException;
+import com.ecommerce.core.usecase.http.ItemOrderHttp;
 import com.ecommerce.core.usecase.http.OrderHttp;
 import com.ecommerce.dataprovider.constants.Constants;
 import com.ecommerce.dataprovider.entity.OrderEntity;
@@ -10,6 +11,7 @@ import com.ecommerce.dataprovider.mapper.ItemOrderMapper;
 import com.ecommerce.dataprovider.mapper.OrderMapper;
 import com.ecommerce.dataprovider.repository.ItemOrderRepository;
 import com.ecommerce.dataprovider.repository.OrderRepository;
+import com.ecommerce.dataprovider.repository.ProductRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -17,6 +19,8 @@ import org.springframework.stereotype.Component;
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
 
 @Slf4j
 @Component
@@ -24,12 +28,15 @@ public class OrderDataProvider implements OrderGateway {
 
     private final OrderRepository orderRepository;
     private final ItemOrderRepository itemOrderRepository;
+    private final ProductDataProvider productDataProvider;
 
     @Autowired
     public OrderDataProvider(final OrderRepository orderRepository,
-                             final ItemOrderRepository itemOrderRepository) {
+                             final ItemOrderRepository itemOrderRepository,
+                             final ProductDataProvider productDataProvider) {
         this.orderRepository = orderRepository;
         this.itemOrderRepository = itemOrderRepository;
+        this.productDataProvider = productDataProvider;
     }
 
     public List<OrderEntity> listAll() {
@@ -45,13 +52,13 @@ public class OrderDataProvider implements OrderGateway {
             orderEntity.get().setItens(itemOrderRepository.listAllByOrderId(id));
             return orderEntity.get();
         } else {
-            log.error(Constants.msgNaoEncontrado);
-            throw new HandlerValidationException(Constants.msgNaoEncontrado);
+            log.error(Constants.msgNotFound);
+            throw new HandlerValidationException(Constants.msgNotFound);
         }
     }
 
     @Transactional
-    public OrderEntity saveOrUpdate(OrderHttp orderHttp, boolean update) {
+    public OrderEntity saveOrUpdate(OrderHttp orderHttp, boolean update) throws HandlerValidationException {
         if (update) {
             return makeOrderUpdate(orderHttp);
         }
@@ -64,14 +71,21 @@ public class OrderDataProvider implements OrderGateway {
         orderRepository.deleteById(id);
     }
 
-    private final OrderEntity makeOrderSave(OrderHttp orderHttp) {
+    private final OrderEntity makeOrderSave(OrderHttp orderHttp) throws HandlerValidationException {
+        if(!verifyQtdProductToSell(orderHttp.getItens())) {
+            throw new HandlerValidationException(Constants.msgOrderNotConfirmed);
+        }
+        verifyQtdProductToSell(orderHttp.getItens());
         OrderEntity orderSaved = orderRepository.saveAndFlush(OrderMapper.httpToEntity(orderHttp));
         orderHttp.getItens().stream().forEach(item -> item.setIdorder(orderSaved.getId()));
         itemOrderRepository.saveAll(ItemOrderMapper.httpToEntity(orderHttp.getItens()));
         return orderSaved;
     }
 
-    private final OrderEntity makeOrderUpdate(OrderHttp orderHttp){
+    private final OrderEntity makeOrderUpdate(OrderHttp orderHttp) throws HandlerValidationException {
+        if(!verifyQtdProductToSell(orderHttp.getItens())) {
+            throw new HandlerValidationException(Constants.msgOrderNotConfirmed);
+        }
         itemOrderRepository.deleteByOrderId(orderHttp.getId());
         OrderEntity orderEntityUpdated = orderRepository.findById(orderHttp.getId()).get();
         orderEntityUpdated.setCustomer(CustomerMapper.httpToEntity(orderHttp.getCustomer()));
@@ -80,6 +94,21 @@ public class OrderDataProvider implements OrderGateway {
         orderEntityUpdated.setItens(ItemOrderMapper.httpToEntity(orderHttp.getItens()));
         itemOrderRepository.saveAll(ItemOrderMapper.httpToEntity(orderHttp.getItens()));
         return orderRepository.saveAndFlush(orderEntityUpdated);
+    }
+
+    private boolean verifyQtdProductToSell(List<ItemOrderHttp> listItens) {
+        AtomicBoolean confirmedOrder = new AtomicBoolean(true);
+        listItens.stream().forEach(item -> {
+            try {
+                if (!productDataProvider.checkItemAvailableToSell(item.getId(), item.getQuantity())
+                && confirmedOrder.get()) {
+                 confirmedOrder.set(false);
+                }
+            } catch (HandlerValidationException e) {
+                e.printStackTrace();
+            }
+        });
+        return confirmedOrder.get();
     }
 
 }
